@@ -14,7 +14,7 @@
     "babel-eslint": "^6.0.4",
     "babel-loader": "^6.2.4",
     "babel-plugin-transform-runtime": "^6.9.0",
-    "babel-polyfill": "^6.9.1",
+    "babel-polyfill": "^6.13.0",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-2": "^6.5.0",
@@ -30,11 +30,11 @@
     "lodash": "^4.13.1",
     "markdown-it": "^7.0.0",
     "markdown-it-link-target": "^1.0.1",
-    "react": "^15.1.0",
-    "react-dom": "^15.1.0",
+    "react": "^15.3.0",
+    "react-dom": "^15.3.0",
     "react-redux": "^4.4.5",
     "redux": "^3.5.2",
-    "redux-form": "^5.2.5",
+    "redux-form": "^5.3.2",
     "redux-logger": "^2.6.1",
     "redux-thunk": "^2.1.0",
     "webpack": "^1.13.1"
