@@ -1,33 +1,49 @@
 const React = require('react');
 const TextareaAutosize = require('react-textarea-autosize').default;
 const has = require('lodash/has');
+const assign = require('lodash/assign');
 
 const MarkdownSource = function(props) {
 
     const contents = has(props.assignProps, 'value') ?
         (props.assignProps.value || '') : (props.contents || '')
 
+    let etc = {};
+
+    if (props.id) {
+        etc.id = props.id;
+    }
+
+    const assignProps = assign({}, props.assignProps);
+
     return (<TextareaAutosize
 
         style={props.style}
-        // key='textarea'
         useCacheForDOMMeasurements
         minRows={6}
+        rows={6}
         maxRows={10}
-        className='form-input'
-        // id="deck_source"
+        className='textarea'
         placeholder={props.placeholder}
+        // TODO: remove
         // onChange={this.onSourceChange}
-        {...props.assignProps}
+        editable={props.editable}
+        {...assignProps}
+        {...etc}
         value={contents}
+
     />);
 };
 
 MarkdownSource.defaultProps = {
     placeholder: '',
-    style: {},
+    style: {
+        height: '136px'
+    },
     assignProps: {},
-    contents: ''
+    contents: '',
+    editable: false,
+    id: void 0
 };
 
 if(process.env.NODE_ENV !== 'production') {
@@ -37,6 +53,10 @@ if(process.env.NODE_ENV !== 'production') {
         editable: React.PropTypes.bool.isRequired,
         placeholder: React.PropTypes.string.isRequired,
         assignProps: React.PropTypes.object,
+        id: React.PropTypes.oneOfType([
+            React.PropTypes.string,
+            React.PropTypes.oneOf([void 0])
+        ])
     };
 }
 
